chore: move function, add basic regtest
