+++ .github/workflows/tests.yml
@@ -15,6 +15,8 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version:  ${{ matrix.python-version }}
+    - name: Install libcgal-dev
+        run: sudo apt-get install -y libcgal-dev
       - name: Install dependencies
         run: |
           pip install pre-commit
