+++ src/moltda/structure_to_vectorization.py
@@ -7,7 +7,7 @@ from .construct_pd import construct_pds
 from .read_file import read_data
 from .vectorize_pds import PersImage, diagrams_to_arrays
 
-__all__ = ["structure_to_pd", "pd_vectorization"]
+__all__ = ["structure_to_pd"]
 
 
 def structure_to_pd(filename: Union[str, Path], supercell_size):
@@ -30,42 +30,3 @@ def structure_to_pd(filename: Union[str, Path], supercell_size):
 
     arr_dgms = diagrams_to_arrays(dgms)  # convert to array representations
     return arr_dgms
-
-
-def pd_vectorization(dgm, spread, weighting, pixels):
-    """
-    Convert persistence diagram array to a vectorized representation.
-
-    Optional: Can add custom specs for scaling the persistence image.
-
-    Args:
-        dgm: Array containing (b, d) points of a persistence diagram.
-        spread: Gaussian spread.
-        weighting: Scheme for weighting points in the persistence diagram.
-        pixels: Pixel size of returned persistence image, e.g. [50, 50]
-
-    Return:
-        Vectorized representation of a persistence diagram, can be used in
-        downstream tasks like machine learning, etc.
-    """
-
-    pim = PersImage(spread=spread, pixels=pixels, weighting_type=weighting, verbose=False)
-
-    image = pim.transform([(x["birth"], x["death"]) for x in dgm])
-
-    return image  # vectorized persistence image
-
-
-def main(filename, supercell_size=None, spread=0.15, weighting="identity", pixels=[50, 50]):
-    """Structure file to vectorized persistence diagram presentation.
-
-    Currently returns a list with the 1D and 2D persistence diagrams of a structure.
-    """
-
-    arr_dgms = structure_to_pd(filename, supercell_size)
-    images = []
-    for dim in [1, 2]:
-        dgm = arr_dgms[f"dim{dim}"]
-        images.append(pd_vectorization(dgm, spread, weighting, pixels))
-
-    return images
